refactor: update paths, add index for enums
