+++ src/client/components/index.ts
@@ -4,3 +4,4 @@ export { GameScreen } from './game-screen/game-screen';
 export { GameEndScreen } from './game-end-screen/game-end-screen';
 export { SettingsScreen } from './settings-screen/settings-screen';
 export { TutorialScreen } from './tutorial-screen/tutorial-screen';
+export { BaseComponent } from './base-component';
